Hide queue operations behind a lock
