+++ vulkano/src/device/mod.rs
@@ -106,16 +106,14 @@ pub(crate) use self::{features::FeaturesFfi, properties::PropertiesFfi};
 pub use self::{
     features::{FeatureRestriction, FeatureRestrictionError, Features},
     properties::Properties,
+    queue::{DebugUtilsError, Queue, QueueFamilyProperties, QueueFlags, QueueGuard},
 };
 use crate::{
     command_buffer::pool::StandardCommandPool,
     descriptor_set::pool::StandardDescriptorPool,
-    instance::{debug::DebugUtilsLabel, Instance},
-    macros::vulkan_bitflags,
+    instance::Instance,
     memory::{pool::StandardMemoryPool, ExternalMemoryHandleType},
-    sync::PipelineStage,
-    OomError, RequirementNotMet, RequiresOneOf, SynchronizedVulkanObject, Version, VulkanError,
-    VulkanObject,
+    OomError, RequirementNotMet, RequiresOneOf, Version, VulkanError, VulkanObject,
 };
 pub use crate::{
     device::extensions::DeviceExtensions,
@@ -123,7 +121,7 @@ pub use crate::{
     fns::DeviceFunctions,
 };
 use ash::vk::Handle;
-use parking_lot::{Mutex, MutexGuard};
+use parking_lot::Mutex;
 use smallvec::SmallVec;
 use std::{
     cell::RefCell,
@@ -143,6 +141,7 @@ pub(crate) mod extensions;
 pub(crate) mod features;
 pub mod physical;
 pub(crate) mod properties;
+mod queue;
 
 /// Represents a Vulkan context.
 #[derive(Debug)]
@@ -442,12 +441,7 @@ impl Device {
                         output.as_mut_ptr(),
                     );
 
-                    Arc::new(Queue {
-                        handle: Mutex::new(output.assume_init()),
-                        device: device.clone(),
-                        queue_family_index,
-                        id,
-                    })
+                    Queue::from_handle(device.clone(), output.assume_init(), queue_family_index, id)
                 },
             )
         };
@@ -700,7 +694,7 @@ impl Device {
     /// of the device (either explicitly or implicitly, for example with a future's destructor)
     /// while this function is waiting.
     ///
-    pub unsafe fn wait(&self) -> Result<(), OomError> {
+    pub unsafe fn wait_idle(&self) -> Result<(), OomError> {
         let fns = self.fns();
         (fns.v1_0.device_wait_idle)(self.handle)
             .result()
@@ -1021,346 +1015,6 @@ impl From<RequirementNotMet> for MemoryFdPropertiesError {
         Self::RequirementNotMet {
             required_for: err.required_for,
             requires_one_of: err.requires_one_of,
-        }
-    }
-}
-
-/// Represents a queue where commands can be submitted.
-// TODO: should use internal synchronization?
-#[derive(Debug)]
-pub struct Queue {
-    handle: Mutex<ash::vk::Queue>,
-    device: Arc<Device>,
-    queue_family_index: u32,
-    id: u32, // id within family
-}
-
-impl Queue {
-    /// Returns the device this queue belongs to.
-    #[inline]
-    pub fn device(&self) -> &Arc<Device> {
-        &self.device
-    }
-
-    /// Returns the queue family that this queue belongs to.
-    #[inline]
-    pub fn queue_family_index(&self) -> u32 {
-        self.queue_family_index
-    }
-
-    /// Returns the index of this queue within its queue family.
-    #[inline]
-    pub fn id_within_family(&self) -> u32 {
-        self.id
-    }
-
-    /// Waits until all work on this queue has finished.
-    ///
-    /// Just like `Device::wait()`, you shouldn't have to call this function in a typical program.
-    #[inline]
-    pub fn wait(&self) -> Result<(), OomError> {
-        unsafe {
-            let fns = self.device.fns();
-            let handle = self.handle.lock();
-            (fns.v1_0.queue_wait_idle)(*handle)
-                .result()
-                .map_err(VulkanError::from)?;
-            Ok(())
-        }
-    }
-
-    /// Opens a queue debug label region.
-    ///
-    /// The [`ext_debug_utils`](crate::instance::InstanceExtensions::ext_debug_utils) must be
-    /// enabled on the instance.
-    #[inline]
-    pub fn begin_debug_utils_label(
-        &self,
-        mut label_info: DebugUtilsLabel,
-    ) -> Result<(), DebugUtilsError> {
-        self.validate_begin_debug_utils_label(&mut label_info)?;
-
-        let DebugUtilsLabel {
-            label_name,
-            color,
-            _ne: _,
-        } = label_info;
-
-        let label_name_vk = CString::new(label_name.as_str()).unwrap();
-        let label_info = ash::vk::DebugUtilsLabelEXT {
-            p_label_name: label_name_vk.as_ptr(),
-            color,
-            ..Default::default()
-        };
-
-        unsafe {
-            let fns = self.device.instance().fns();
-            let handle = self.handle.lock();
-            (fns.ext_debug_utils.queue_begin_debug_utils_label_ext)(*handle, &label_info);
-        }
-
-        Ok(())
-    }
-
-    fn validate_begin_debug_utils_label(
-        &self,
-        _label_info: &mut DebugUtilsLabel,
-    ) -> Result<(), DebugUtilsError> {
-        if !self
-            .device()
-            .instance()
-            .enabled_extensions()
-            .ext_debug_utils
-        {
-            return Err(DebugUtilsError::RequirementNotMet {
-                required_for: "`begin_debug_utils_label`",
-                requires_one_of: RequiresOneOf {
-                    instance_extensions: &["ext_debug_utils"],
-                    ..Default::default()
-                },
-            });
-        }
-
-        Ok(())
-    }
-
-    /// Closes a queue debug label region.
-    ///
-    /// The [`ext_debug_utils`](crate::instance::InstanceExtensions::ext_debug_utils) must be
-    /// enabled on the instance.
-    ///
-    /// # Safety
-    ///
-    /// - There must be an outstanding queue label region begun with `begin_debug_utils_label` in
-    ///   the queue.
-    #[inline]
-    pub unsafe fn end_debug_utils_label(&self) -> Result<(), DebugUtilsError> {
-        self.validate_end_debug_utils_label()?;
-
-        {
-            let fns = self.device.instance().fns();
-            let handle = self.handle.lock();
-            (fns.ext_debug_utils.queue_end_debug_utils_label_ext)(*handle);
-        }
-
-        Ok(())
-    }
-
-    fn validate_end_debug_utils_label(&self) -> Result<(), DebugUtilsError> {
-        if !self
-            .device()
-            .instance()
-            .enabled_extensions()
-            .ext_debug_utils
-        {
-            return Err(DebugUtilsError::RequirementNotMet {
-                required_for: "`end_debug_utils_label`",
-                requires_one_of: RequiresOneOf {
-                    instance_extensions: &["ext_debug_utils"],
-                    ..Default::default()
-                },
-            });
-        }
-
-        // VUID-vkQueueEndDebugUtilsLabelEXT-None-01911
-        // TODO: not checked, so unsafe for now
-
-        Ok(())
-    }
-
-    /// Inserts a queue debug label.
-    ///
-    /// The [`ext_debug_utils`](crate::instance::InstanceExtensions::ext_debug_utils) must be
-    /// enabled on the instance.
-    #[inline]
-    pub fn insert_debug_utils_label(
-        &mut self,
-        mut label_info: DebugUtilsLabel,
-    ) -> Result<(), DebugUtilsError> {
-        self.validate_insert_debug_utils_label(&mut label_info)?;
-
-        let DebugUtilsLabel {
-            label_name,
-            color,
-            _ne: _,
-        } = label_info;
-
-        let label_name_vk = CString::new(label_name.as_str()).unwrap();
-        let label_info = ash::vk::DebugUtilsLabelEXT {
-            p_label_name: label_name_vk.as_ptr(),
-            color,
-            ..Default::default()
-        };
-
-        unsafe {
-            let fns = self.device.instance().fns();
-            let handle = self.handle.lock();
-            (fns.ext_debug_utils.queue_insert_debug_utils_label_ext)(*handle, &label_info);
-        }
-
-        Ok(())
-    }
-
-    fn validate_insert_debug_utils_label(
-        &self,
-        _label_info: &mut DebugUtilsLabel,
-    ) -> Result<(), DebugUtilsError> {
-        if !self
-            .device()
-            .instance()
-            .enabled_extensions()
-            .ext_debug_utils
-        {
-            return Err(DebugUtilsError::RequirementNotMet {
-                required_for: "`insert_debug_utils_label`",
-                requires_one_of: RequiresOneOf {
-                    instance_extensions: &["ext_debug_utils"],
-                    ..Default::default()
-                },
-            });
-        }
-
-        Ok(())
-    }
-}
-
-unsafe impl SynchronizedVulkanObject for Queue {
-    type Object = ash::vk::Queue;
-
-    #[inline]
-    fn internal_object_guard(&self) -> MutexGuard<Self::Object> {
-        self.handle.lock()
-    }
-}
-
-unsafe impl DeviceOwned for Queue {
-    fn device(&self) -> &Arc<Device> {
-        &self.device
-    }
-}
-
-impl PartialEq for Queue {
-    fn eq(&self, other: &Self) -> bool {
-        self.id == other.id
-            && self.queue_family_index == other.queue_family_index
-            && self.device == other.device
-    }
-}
-
-impl Eq for Queue {}
-
-impl Hash for Queue {
-    #[inline]
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.id.hash(state);
-        self.queue_family_index.hash(state);
-        self.device.hash(state);
-    }
-}
-
-/// Properties of a queue family in a physical device.
-#[derive(Clone, Debug)]
-#[non_exhaustive]
-pub struct QueueFamilyProperties {
-    /// Attributes of the queue family.
-    pub queue_flags: QueueFlags,
-
-    /// The number of queues available in this family.
-    ///
-    /// This guaranteed to be at least 1 (or else that family wouldn't exist).
-    pub queue_count: u32,
-
-    /// If timestamps are supported, the number of bits supported by timestamp operations.
-    /// The returned value will be in the range 36..64.
-    ///
-    /// If timestamps are not supported, this is `None`.
-    pub timestamp_valid_bits: Option<u32>,
-
-    /// The minimum granularity supported for image transfers, in terms of `[width, height, depth]`.
-    pub min_image_transfer_granularity: [u32; 3],
-}
-
-impl QueueFamilyProperties {
-    /// Returns whether the queues of this family support a particular pipeline stage.
-    #[inline]
-    pub fn supports_stage(&self, stage: PipelineStage) -> bool {
-        ash::vk::QueueFlags::from(self.queue_flags).contains(stage.required_queue_flags())
-    }
-}
-
-impl From<ash::vk::QueueFamilyProperties> for QueueFamilyProperties {
-    #[inline]
-    fn from(val: ash::vk::QueueFamilyProperties) -> Self {
-        Self {
-            queue_flags: val.queue_flags.into(),
-            queue_count: val.queue_count,
-            timestamp_valid_bits: (val.timestamp_valid_bits != 0)
-                .then_some(val.timestamp_valid_bits),
-            min_image_transfer_granularity: [
-                val.min_image_transfer_granularity.width,
-                val.min_image_transfer_granularity.height,
-                val.min_image_transfer_granularity.depth,
-            ],
-        }
-    }
-}
-
-vulkan_bitflags! {
-    /// Attributes of a queue or queue family.
-    #[non_exhaustive]
-    QueueFlags = QueueFlags(u32);
-
-    /// Queues of this family can execute graphics operations.
-    graphics = GRAPHICS,
-
-    /// Queues of this family can execute compute operations.
-    compute = COMPUTE,
-
-    /// Queues of this family can execute transfer operations.
-    transfer = TRANSFER,
-
-    /// Queues of this family can execute sparse memory management operations.
-    sparse_binding = SPARSE_BINDING,
-
-    /// Queues of this family can be created using the `protected` flag.
-    protected = PROTECTED {
-        api_version: V1_1,
-    },
-
-    /// Queues of this family can execute video decode operations.
-    video_decode = VIDEO_DECODE_KHR {
-        device_extensions: [khr_video_decode_queue],
-    },
-
-    /// Queues of this family can execute video encode operations.
-    video_encode = VIDEO_ENCODE_KHR {
-        device_extensions: [khr_video_encode_queue],
-    },
-}
-
-/// Error that can happen when submitting a debug utils command to a queue.
-#[derive(Clone, Debug)]
-pub enum DebugUtilsError {
-    RequirementNotMet {
-        required_for: &'static str,
-        requires_one_of: RequiresOneOf,
-    },
-}
-
-impl Error for DebugUtilsError {}
-
-impl Display for DebugUtilsError {
-    #[inline]
-    fn fmt(&self, f: &mut Formatter) -> Result<(), FmtError> {
-        match self {
-            Self::RequirementNotMet {
-                required_for,
-                requires_one_of,
-            } => write!(
-                f,
-                "a requirement was not met for: {}; requires one of: {}",
-                required_for, requires_one_of,
-            ),
         }
     }
 }
