+++ vulkano/src/sync/future/semaphore_signal.rs
@@ -191,7 +191,7 @@ where
                 // TODO: handle errors?
                 self.flush().unwrap();
                 // Block until the queue finished.
-                self.queue().unwrap().wait().unwrap();
+                self.queue().unwrap().lock().wait_idle().unwrap();
                 self.previous.signal_finished();
             }
         }
